data/admin/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
Flask-WTF==1.1.1
WTForms==3.0.1
python-dotenv==1.0.0 